Pre-Alpha version of tuning workbench synth
The tuning workbench synth is a VI which lets you edit KBM and
SCL in place to see how they sound. It is mostly pedagogical.

This checkin is the earliest possible alpha I could share. It makes
sound, respects tuning, and has the UI, although many parts of hte
UI and synth are incomplete.